+++ tools/results_uploader/src/resultstore_client.py
@@ -17,8 +17,10 @@
 """Resultstore client for Mobly tests."""
 
 import datetime
+import enum
 import logging
 import posixpath
+import urllib.parse
 import uuid
 
 from google.auth import credentials
@@ -27,16 +29,15 @@ from googleapiclient import discovery
 import httplib2
 
 _DEFAULT_CONFIGURATION = 'default'
-_RESULTSTORE_BASE_LINK = 'https://source.cloud.google.com/results/invocations'
-
-_STATUS_PASSED = 'PASSED'
-_STATUS_FAILED = 'FAILED'
-_STATUS_UNKNOWN = 'UNKNOWN'
+_RESULTSTORE_BASE_LINK = 'https://btx.cloud.google.com/invocations'
 
 
-def _get_status(passed: bool) -> str:
-    """Get the corresponding status depending on if the test run passed."""
-    return _STATUS_PASSED if passed else _STATUS_FAILED
+class Status(enum.Enum):
+    """Aggregate status of the Resultstore invocation and target."""
+    PASSED = 'PASSED'
+    FAILED = 'FAILED'
+    SKIPPED = 'SKIPPED'
+    UNKNOWN = 'UNKNOWN'
 
 
 class ResultstoreClient:
@@ -65,12 +66,34 @@ class ResultstoreClient:
         self._invocation_id = ''
         self._authorization_token = ''
         self._target_id = ''
+        self._encoded_target_id = ''
 
-        self._status = _STATUS_UNKNOWN
+        self._status = Status.UNKNOWN
 
-    def set_status(self, passed: bool) -> None:
-        """Sets the status depending on if the test run passed."""
-        self._status = _get_status(passed)
+    @property
+    def _invocation_name(self):
+        """The resource name for the invocation."""
+        if not self._invocation_id:
+            return ''
+        return f'invocations/{self._invocation_id}'
+
+    @property
+    def _target_name(self):
+        """The resource name for the target."""
+        if not (self._invocation_name or self._encoded_target_id):
+            return ''
+        return f'{self._invocation_name}/targets/{self._encoded_target_id}'
+
+    @property
+    def _configured_target_name(self):
+        """The resource name for the configured target."""
+        if not self._target_name:
+            return
+        return f'{self._target_name}/configuredTargets/{_DEFAULT_CONFIGURATION}'
+
+    def set_status(self, status: Status) -> None:
+        """Sets the overall test run status."""
+        self._status = status
 
     def create_invocation(self) -> str:
         """Creates an invocation.
@@ -134,8 +157,7 @@ class ResultstoreClient:
         Returns:
           The target ID.
         """
-        parent = f'invocations/{self._invocation_id}'
-        logging.info('creating target in %s...', parent)
+        logging.info('creating target in %s...', self._invocation_name)
         if self._target_id:
             logging.warning(
                 'target %s already exists, skipping creation...',
@@ -143,6 +165,7 @@ class ResultstoreClient:
             )
             return
         self._target_id = target_id or str(uuid.uuid4())
+        self._encoded_target_id = urllib.parse.quote(self._target_id, safe='')
         target = {
             'id': {
                 'invocationId': self._invocation_id,
@@ -155,7 +178,7 @@ class ResultstoreClient:
             .targets()
             .create(
                 body=target,
-                parent=parent,
+                parent=self._invocation_name,
                 targetId=self._target_id,
                 authorizationToken=self._authorization_token,
             )
@@ -166,8 +189,7 @@ class ResultstoreClient:
 
     def create_configured_target(self) -> None:
         """Creates a configured target."""
-        parent = f'invocations/{self._invocation_id}/targets/{self._target_id}'
-        logging.info('creating configured target in %s...', parent)
+        logging.info('creating configured target in %s...', self._target_name)
         configured_target = {
             'id': {
                 'invocationId': self._invocation_id,
@@ -181,7 +203,7 @@ class ResultstoreClient:
             .configuredTargets()
             .create(
                 body=configured_target,
-                parent=parent,
+                parent=self._target_name,
                 configId=_DEFAULT_CONFIGURATION,
                 authorizationToken=self._authorization_token,
             )
@@ -199,11 +221,7 @@ class ResultstoreClient:
         Returns:
           The action ID.
         """
-        parent = (
-            f'invocations/{self._invocation_id}/targets/{self._target_id}/'
-            'configuredTargets/default'
-        )
-        logging.info('creating action in %s...', parent)
+        logging.info('creating action in %s...', self._configured_target_name)
         action_id = str(uuid.uuid4())
         files = [
             {'uid': path, 'uri': posixpath.join(gcs_path, path)}
@@ -226,7 +244,7 @@ class ResultstoreClient:
             .actions()
             .create(
                 body=action,
-                parent=parent,
+                parent=self._configured_target_name,
                 actionId=action_id,
                 authorizationToken=self._authorization_token,
             )
@@ -239,14 +257,11 @@ class ResultstoreClient:
 
     def merge_configured_target(self):
         """Merges a configured target."""
-        name = (
-            f'invocations/{self._invocation_id}/targets/{self._target_id}/'
-            'configuredTargets/default'
-        )
-        logging.info('merging configured target %s...', name)
+        logging.info('merging configured target %s...',
+                     self._configured_target_name)
         merge_request = {
             'configuredTarget': {
-                'statusAttributes': {'status': self._status},
+                'statusAttributes': {'status': self._status.value},
             },
             'authorizationToken': self._authorization_token,
             'updateMask': 'statusAttributes',
@@ -257,7 +272,7 @@ class ResultstoreClient:
             .configuredTargets()
             .merge(
                 body=merge_request,
-                name=name,
+                name=self._configured_target_name,
             )
         )
         res = request.execute(http=self._http)
@@ -265,11 +280,8 @@ class ResultstoreClient:
 
     def finalize_configured_target(self):
         """Finalizes a configured target."""
-        name = (
-            f'invocations/{self._invocation_id}/targets/{self._target_id}/'
-            'configuredTargets/default'
-        )
-        logging.info('finalizing configured target %s...', name)
+        logging.info('finalizing configured target %s...',
+                     self._configured_target_name)
         finalize_request = {
             'authorizationToken': self._authorization_token,
         }
@@ -279,7 +291,7 @@ class ResultstoreClient:
             .configuredTargets()
             .finalize(
                 body=finalize_request,
-                name=name,
+                name=self._configured_target_name,
             )
         )
         res = request.execute(http=self._http)
@@ -287,11 +299,10 @@ class ResultstoreClient:
 
     def merge_target(self):
         """Merges a target."""
-        name = f'invocations/{self._invocation_id}/targets/{self._target_id}'
-        logging.info('merging target %s...', name)
+        logging.info('merging target %s...', self._target_name)
         merge_request = {
             'target': {
-                'statusAttributes': {'status': self._status},
+                'statusAttributes': {'status': self._status.value},
             },
             'authorizationToken': self._authorization_token,
             'updateMask': 'statusAttributes',
@@ -301,7 +312,7 @@ class ResultstoreClient:
             .targets()
             .merge(
                 body=merge_request,
-                name=name,
+                name=self._target_name,
             )
         )
         res = request.execute(http=self._http)
@@ -309,8 +320,7 @@ class ResultstoreClient:
 
     def finalize_target(self):
         """Finalizes a target."""
-        name = f'invocations/{self._invocation_id}/targets/{self._target_id}'
-        logging.info('finalizing target %s...', name)
+        logging.info('finalizing target %s...', self._target_name)
         finalize_request = {
             'authorizationToken': self._authorization_token,
         }
@@ -319,39 +329,35 @@ class ResultstoreClient:
             .targets()
             .finalize(
                 body=finalize_request,
-                name=(
-                    f'invocations/{self._invocation_id}/targets/'
-                    f'{self._target_id}'
-                ),
+                name=self._target_name,
             )
         )
         res = request.execute(http=self._http)
         logging.debug('invocations.targets.finalize: %s', res)
         self._target_id = ''
+        self._encoded_target_id = ''
 
     def merge_invocation(self):
         """Merges an invocation."""
-        name = f'invocations/{self._invocation_id}'
-        logging.info('merging invocation %s...', name)
+        logging.info('merging invocation %s...', self._invocation_name)
         merge_request = {
-            'invocation': {'statusAttributes': {'status': self._status}},
+            'invocation': {'statusAttributes': {'status': self._status.value}},
             'updateMask': 'statusAttributes',
             'authorizationToken': self._authorization_token,
         }
         request = self._service.invocations().merge(body=merge_request,
-                                                    name=name)
+                                                    name=self._invocation_name)
         res = request.execute(http=self._http)
         logging.debug('invocations.merge: %s', res)
 
     def finalize_invocation(self):
         """Finalizes an invocation."""
-        name = f'invocations/{self._invocation_id}'
-        logging.info('finalizing invocation %s...', name)
+        logging.info('finalizing invocation %s...', self._invocation_name)
         finalize_request = {
             'authorizationToken': self._authorization_token,
         }
         request = self._service.invocations().finalize(
-            body=finalize_request, name=name
+            body=finalize_request, name=self._invocation_name
         )
         res = request.execute(http=self._http)
         logging.debug('invocations.finalize: %s', res)